+++ Changelog.md
@@ -1,7 +1,15 @@
 # Release History
 
 ### 1.0.1 ( 2022-09-19)
+- [x] Fix: Deep copy process is losing object properties that are equal to 'null';
+
+
+
+### 1.0.1 ( 2022-09-19)
 - [x] Fix: Deep copy is not working.
+- [ ] Bug: Deep copy process is losing object properties that are equal to 'null';
+
+
 
 ### 1.0.0 (2022-09-18)
  - [x] Initial code;